# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Marine Corps logistics calculator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sync/settings:
    patch:
      operationId: updateSyncSettings
      tags: [sync]
      summary: Update auto-sync configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncSettingsInput"
      responses:
        "200":
          description: Updated sync status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncStatus"
...
  /sync/outbox/{id}:
    delete:
      operationId: dismissOutboxRecord
      tags: [sync]
      summary: Permanently remove a failed outbox record (dismiss it from the queue)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Dismissed
        "404":
          description: Record not found
  /sync/outbox/{id}/retry:
    post:
      operationId: retryOutboxRecord
      tags: [sync]
      summary: Reset a failed outbox record back to pending so it re-syncs immediately
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Record queued for retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncOutboxItem"
        "404":
          description: Record not found
  /sync/last-run:
    get:
      operationId: getSyncLastRun
      tags: [sync]
      summary: Details of the most recent SPIRE sync run
      responses:
        "200":
          description: Last sync run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncLastRun"
        "204":
          description: No sync run has been performed yet
...
components:
  schemas:
...
    SyncStatus:
      type: object
      properties:
        connected:
          type: boolean
        upstreamSystem:
          type: string
        lastSyncAt:
          type: string
          format: date-time
          nullable: true
        pendingChanges:
          type: integer
        latencyMs:
          type: integer
          nullable: true
        autoSyncEnabled:
          type: boolean
        autoSyncIntervalMinutes:
          type: integer
      required: [connected, upstreamSystem, pendingChanges, autoSyncEnabled, autoSyncIntervalMinutes]
    SyncSettingsInput:
      type: object
      properties:
        autoSyncEnabled:
          type: boolean
        autoSyncIntervalMinutes:
          type: integer
          minimum: 1
          maximum: 60
    SyncOutboxItem:
      type: object
      properties:
        id:
          type: string
        entityKind:
          type: string
        entityId:
          type: string
        unitId:
          type: string
          nullable: true
        op:
          type: string
        createdAt:
          type: string
          format: date-time
        status:
          type: string
        lastError:
          type: string
          nullable: true
        itemName:
          type: string
          nullable: true
        unitName:
          type: string
          nullable: true
      required: [id, entityKind, entityId, op, createdAt, status]
    SyncPushResult:
      type: object
      properties:
        outboxId:
          type: string
        entityKind:
          type: string
        entityId:
          type: string
        success:
          type: boolean
        error:
          type: string
          nullable: true
        unitName:
          type: string
          nullable: true
        itemName:
          type: string
          nullable: true
      required: [outboxId, entityKind, entityId, success]
    SyncLastRun:
      type: object
      properties:
        id:
          type: string
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        pushedCount:
          type: integer
        failedCount:
          type: integer
        catalogMatched:
          type: integer
        catalogNew:
          type: integer
        catalogChanged:
          type: integer
        latencyMs:
          type: integer
          nullable: true
        results:
          type: array
          items:
            $ref: "#/components/schemas/SyncPushResult"
      required: [id, startedAt, pushedCount, failedCount, catalogMatched, catalogNew, catalogChanged]
...